requirements.txt
# requirements.txt

# Core dependencies
python==3.9.16  # Ensure compatibility with the project

# AI and Machine Learning
transformers==4.33.2  # For pre-trained language models
torch==2.0.1  # PyTorch for deep learning
scikit-learn==1.3.1  # For machine learning utilities
numpy==1.25.2  # Numerical computations
pandas==2.1.1  # Data manipulation and analysis
nltk==3.9.1  # Natural Language Toolkit for text processing
spacy==3.6.1  # NLP library for tokenization and named entity recognition
sentence-transformers==2.2.2  # For sentence embeddings

# Web scraping and APIs
beautifulsoup4==4.12.2  # For web scraping
requests==2.31.0  # For making HTTP requests
selenium==4.12.0  # For browser automation (if needed for scraping dynamic content)

# Data visualization
matplotlib==3.8.0  # For plotting and visualizations
seaborn==0.12.2  # Statistical data visualization

# Backend and API
fastapi==0.100.1  # For building the backend API
uvicorn==0.23.2  # ASGI server for FastAPI
pydantic==2.3.0  # Data validation and settings management

# Database
sqlalchemy==2.1.1  # Database ORM
psycopg2-binary==2.9.8  # PostgreSQL database adapter

# Authentication and Security
python-jose==3.3.0  # JWT handling
passlib[bcrypt]==1.7.4  # Password hashing

# Testing
pytest==7.4.2  # For testing
pytest-cov==4.1.0  # For test coverage reporting

# Linting and Code Quality
flake8==6.1.0  # For linting
black==23.9.1  # Code formatter
isort==5.12.0  # Import sorting

# Docker and Deployment
docker==6.1.3  # Docker SDK for Python
gunicorn==21.2.0  # WSGI HTTP server for production

# Miscellaneous
python-dotenv==1.0.0  # For managing environment variables
tqdm==4.66.1  # Progress bar utility
loguru==0.7.1  # Advanced logging